Updates column plot.

+++ python/magnetizer/visualization.py
@@ -31,6 +31,10 @@ from scipy.interpolate import UnivariateSpline
 
 units_dict = {}
 
+
+mnras_column_size = 3.32153
+mnras_text_size = 6.97522
+
 default_quantities = [
                         'V',
                         'Omega',
@@ -446,46 +450,83 @@ def generate_portfolio(run_obj, selected_quantities=None, binning_obj=None,
     return figures
 
 
-def plot_time_evolution_column(t, name, keypos=None, limits=None, fig=None,
-                               t_ticks=None, z_ticks=None, color='b'):
+def plot_evolution_column(t, name, keypos=None, limits=None, x_limits=None,
+                          t_ticks=None, z_ticks=None, use_z=False,
+                          fig=None, figsize=None, color='b'):
     if fig is None:
         first = True
-        fig, ax = plt.subplots(t.nbins, sharex=True,
-                            figsize=(mnras_text_size/4.1,mnras_text_size*0.55))
+        if figsize is None:
+            figsize=(mnras_text_size/4.1,mnras_text_size*0.55)
+        fig, ax = plt.subplots(t.nbins, sharex=True, figsize=figsize)
     else:
         first = False
         ax = fig.axes
-    zs_or_ts = t.times.value
+    if not use_z:
+        zs_or_ts = t.times.value
 
-    t_limits = (t.times.min()/u.Gyr, t.times.max()/u.Gyr)
+        if x_limits is None:
+            x_limits = (t.times.min()/u.Gyr, t.times.max()/u.Gyr)
 
-    if t_ticks is None:
-        ltimes = np.linspace(t.times.max()/u.Gyr, t.times.min()/u.Gyr, 5)
+        if t_ticks is None:
+            t_ticks = np.linspace(t.times.max()/u.Gyr, t.times.min()/u.Gyr, 5)
+        else:
+            t_ticks = np.array(t_ticks)
+
+        if z_ticks is not None:
+            # Find corresponding redshifts with spline interpolation
+            z_to_t_spline_converter = UnivariateSpline(t.run.redshifts[::-1],
+                                                      t.run.times[::-1]/u.Gyr,
+                                                      k=5, s=0)
+
+            z_tick_pos = z_to_t_spline_converter(z_ticks)
+        else:
+            z_tick_pos = t_ticks
+            t_to_z_spline_converter = UnivariateSpline(t.run.times/u.Gyr,
+                                                      t.run.redshifts, k=5, s=0)
+
+            z_ticks = t_to_z_spline_converter(z_tick_pos)
+
+        tlabels = ['{0:.1f}'.format(x) for x in t_ticks]
+        zlabels = ['{0:.1f}'.format(abs(x)) for x in z_ticks]
     else:
-        ltimes = np.array(t_ticks)
+        zs_or_ts = t.zs
 
-    if z_ticks is not None:
-        # Find corresponding redshifts with spline interpolation
-        z_to_t_spline_converter = UnivariateSpline(t.run.redshifts[::-1],
-                                                   t.run.times[::-1]/u.Gyr,
-                                                   k=5, s=0)
+        if x_limits is None:
+            x_limits = (t.zs.min(), t.zs.max())
 
-        z_tick_pos = z_to_t_spline_converter(z_ticks)
-    else:
-        z_tick_pos = ltimes
-        t_to_z_spline_converter = UnivariateSpline(t.run.times/u.Gyr,
-                                                   t.run.redshifts,
-                                                   k=5, s=0)
-        z_ticks = t_to_z_spline_converter(z_tick_pos)
+        if z_ticks is None:
+            z_ticks = np.linspace(t.zs.min(), t.zs.max(), 5)
+        else:
+            z_ticks = np.array(z_ticks)
 
-    tlabels = ['{0:.1f}'.format(x) for x in ltimes]
-    zlabels = ['{0:.1f}'.format(abs(x)) for x in z_ticks]
+        if t_ticks is not None:
+            # Find corresponding redshifts with spline interpolation
+            t_to_z_spline_converter = UnivariateSpline(t.run.times/u.Gyr,
+                                                      t.run.redshifts, k=5, s=0)
+
+            t_tick_pos = t_to_z_spline_converter(t_ticks)
+        else:
+            t_tick_pos = z_ticks
+
+            z_to_t_spline_converter = UnivariateSpline(t.run.redshifts[::-1],
+                                                      t.run.times[::-1]/u.Gyr,
+                                                      k=5, s=0)
+            t_ticks = z_to_t_spline_converter(t_tick_pos)
+
+        tlabels = ['{0:.1f}'.format(x) for x in t_ticks]
+        zlabels = ['{0:.1f}'.format(abs(x)) for x in z_ticks]
+
 
     if limits is None:
         limits = (np.nanmin(t.lower), np.nanmax(t.upper))
-    for i, b in enumerate(t.bins.bins):
+    try:
+        bins = t.bins.bins
+    except:
+        bins = t.bins[0].bins
+
+    for i, b in enumerate(bins):
         if keypos is not None:
-            ax[i].text(keypos[0],keypos[1],format_log_mass(b), size=6)
+            ax[i].text(keypos[0],keypos[1],format_log_mass(b), size=5.85)
 
         ax[i].plot(zs_or_ts, t.med[i], color=color)
         ax[i].plot(zs_or_ts, t.lower[i], color=color, linestyle=':')
@@ -498,18 +539,28 @@ def plot_time_evolution_column(t, name, keypos=None, limits=None, fig=None,
             ax[i].set_ylim(limits)
             ax[i].set_yticks(ax[i].get_yticks()[1:-1])
     if first:
-        ax[0].set_xlim(*t_limits)
-        ax[0].xaxis.set_ticks(ltimes)
-        ax[0].xaxis.set_ticklabels(tlabels)
-        ax[-1].set_xlabel(r"$t\;[\rm Gyr]$")
-
+        ax[0].set_xlim(*x_limits)
+        if not use_z:
+            ax[0].xaxis.set_ticks(t_ticks)
+            ax[0].xaxis.set_ticklabels(tlabels)
+            ax[-1].set_xlabel(r"$t\;[\rm Gyr]$")
+        else:
+            ax[0].xaxis.set_ticks(z_ticks)
+            ax[0].xaxis.set_ticklabels(zlabels)
+            ax[-1].set_xlabel(r"$z$")
 
         ax2 = ax[0].twiny()
 
-        ax2.set_xlim(*t_limits)
-        ax2.xaxis.set_ticks(z_tick_pos)
-        ax2.xaxis.set_ticklabels(zlabels)
-        ax2.set_xlabel(r"$z$")
+        if not use_z:
+            ax2.set_xlim(*x_limits)
+            ax2.xaxis.set_ticks(z_tick_pos)
+            ax2.xaxis.set_ticklabels(zlabels)
+            ax2.set_xlabel(r"$z$")
+        else:
+            ax2.set_xlim(*x_limits)
+            ax2.xaxis.set_ticks(t_tick_pos)
+            ax2.xaxis.set_ticklabels(tlabels)
+            ax2.set_xlabel(r"$t\;[\rm Gyr]$")
 
     fig.subplots_adjust(hspace=0,
                         left=0.28,
@@ -517,7 +568,6 @@ def plot_time_evolution_column(t, name, keypos=None, limits=None, fig=None,
                         top=0.92,
                         bottom=0.1)
     return fig
-
 
 def plot_redshift_evolution(quantity, mag_run, position=None,
                             target_redshifts=None, bin_objs=None,
